Create AkashG.Activity1
Case Study : Online Shopping System  
Problem Statement  
Design a system to input and display the details of products in a shopping cart. 
The program should use an array of Product objects to hold product details and apply a copy constructor to duplicate a Product object. By utilizing the copy constructor, you can easily create a new Product object by copying the attributes from an existing one, which is useful for creating similar products in the cart